Present

Description

Make presentations for data types.

Documentation

data Presentation

A presentation of a level of a data type.

Constructors

Integer !Text !Text

Floating !Text !Text

Char !Text !Text

Alg !Text !Text ![(Text, ID)]

Tuple !Text ![(Text, ID)]

List !Text ![(Text, ID)]

String !Text ![(Text, ID)]

data ID

present :: Data a => ID -> a -> Maybe Presentation

Present the breadth-first level of a data type.

asData :: Data a => a -> a

A helpful function for editors to force that a value is an instance of Data, before we actually start using it.